Dynamic Queries Java Classes
This program is able to receive a JSON object from Postman (firstName, lastName, etc.), and is able to conduct a search based on the criteria given within the object. I have connected my local mySQL database to my program through the application.properties file, in which you must do the same!